+++ client/timing/timing.go
@@ -11,9 +11,9 @@ const (
 	t0 = 1658429955
 )
 
-func RoundIDForTimestamp(t uint64) uint64 {
+func RoundIDForTimestamp(t uint64) uint32 {
 
-	roundID := uint64((t - t0 + 30) / 90)
+	roundID := uint32((t - t0 + 30) / 90)
 
 	return roundID
 }
@@ -51,7 +51,7 @@ func NextChoosePhaseEndPointers(t uint64) (*int, *uint64) {
 func LastCollectPhaseStart(t uint64) (int, uint64) {
 	roundID := RoundIDForTimestamp(t)
 
-	startTimestamp := t0 + roundID*90 - 30
+	startTimestamp := t0 + uint64(roundID)*90 - 30
 
 	return int(roundID), startTimestamp
 }
